pyproject.toml
[tool.poetry]
name = "cognite_synthetic_tags"
version = "0.5.1"
description = "Framework for working easily with tags, synthetic and regular, from Cognite Data Fusion (CDF) API."
authors = ["Fran Hrzenjak <fran.hrzenjak@cognite.com>"]
license = "Apache License 2.0"

[tool.black]
line-length = 80

[tool.isort]
line_length=80
multi_line_output=3
skip_glob = '^((?!py$).)*$'    # this makes sort all Python files
include_trailing_comma=true
force_grid_wrap=0
use_parentheses=true
ensure_newline_before_comments=true

[tool.coverage.run]
branch=true
source=["cognite_synthetic_tags"]

[tool.poetry.dependencies]
python = ">=3.9,<4"
cognite-sdk = {version = ">7,<8", extras = ["pandas"]}

[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
pre-commit = "^3.6.2"
pytest = "^8.0.2"
pytest-mock = "^3.12.0"
coverage = {version = "^7.4.3", extras = ["toml"]}
python-dotenv = "^1.0.1"
pytest-regressions = "^2.5.0"
arrow = "^1.3.0"
pytest-cov = "^5.0.0"

[build-system]
requires = [
    "poetry-core>=1.0.0",
    "setuptools>60.0",
    "wheel",
]
build-backend = "poetry.core.masonry.api"
